Fix resource area max default (#160)
Default was ``False`` with the intention to deactivate the constraint. Quite the opposite actually happened: ``False`` was evaluated to ``0`` _activating_ this constraint to its maximum.

Now the default is ``inf`` which will actually deactivate the constraint (in compliance with other upper bound constraints).

In the documentation, resource areas were related to collector areas. But these constraints can equally be used for all sorts of other technologies with area restrictions.Hence, I made the documentation slightly more generic.